move file inside package
